Fixed item rating

+++ src/components/Item.js
@@ -17,7 +17,7 @@ function Item(props) {
                         <div className="card-body cardview">
                             <Link to="" style={{'color': 'black'}}><h5 className="card-title">{ props.item_name }</h5></Link>
                             <h5 className="my-2">Rs. {props.price} /-</h5>
-                            <Rating name="half-rating" value={props.rating.intValue} precision={0.5} readOnly />
+                            <Rating name="half-rating" value={props.rating} precision={0.5} readOnly />
                             <hr />
                             <Button className="px-4 ms-1 me-2" variant="success"><BoltIcon sx={{ fontSize: 18 }} className="me-2" />Buy now</Button>{' '}
                             <Button className="px-4" variant="primary"><ShoppingCartIcon sx={{ fontSize: 18 }} className="me-2" />Add to cart</Button>{' '}
